adventure_creator/requirements.txt
google-adk>=0.5.0
google-cloud-secret-manager>=2.16.0
requests>=2.25.0